Full-ish Release
Pretty much everything is in working order now.

During the creation of the meme it will make sure that the text fill at
least 15% of the image at the top and bottom. This will hopefully
prevent from some images having text that is too small to read.

Newline characters are added a little more frequently to make sure that
links wrap around the edge of the picture, instead of taking up an
entire line and messing up the proportions

A “no text zone” was added to the middle of the image. The picture is
divided into 9ths and the middle portion is reserved so that text will
not overflow. It is 11% of the image. Text will scale down
appropriately to mind this gap